+++ src/meta.jl
@@ -48,22 +48,20 @@ function outer_constructor(obj::DataType)
 end
 =#
 
-function outer_constructor(obj::DataType, defaults::Tuple{Vararg{Real}})
+function generate_outer_constructor(obj::DataType, defaults::Tuple{Vararg{Real}})
     fields = fieldnames(obj)
     length(defaults) == length(fields) || error("wrong size")
     parameters  = [param.name => param.ub for param in obj.parameters]  # [:T=>Float64, :U=>Int]
-    #println(parameters)
     fieldparams = Symbol[fieldtype(obj, field).parameters[1].name for field in fields]
-    #println(fieldparams)
     first_idx   = [param => findfirst(fieldparams, param) for param in keys(parameters)]
-    #println(first_idx)
+
     defaultparams = Symbol[if i == first_idx[fieldparams[i]]
                         get_default(parameters[fieldparams[i]]).name.name
                     else
                         fieldparams[i]
                     end for i in eachindex(fields)]
 
-    arguments = [Expr(:(=), :($(fields[i])::Argument{$(fieldparams[i])}),
+    arguments = [Expr(:kw, :($(fields[i])::Argument{$(fieldparams[i])}),
                             :(convert($(defaultparams[i]), $(defaults[i]))))
                  for i in eachindex(fields)]
 
@@ -71,9 +69,19 @@ function outer_constructor(obj::DataType, defaults::Tuple{Vararg{Real}})
                        [Expr(:(<:), param.name, param.ub.name.name) for param in obj.parameters]...)
 
     definition_ls = Expr(:call, constructor, arguments...)
+
+    definition_rs = Expr(:call, 
+                         Expr(:curly, obj.name.name, 
+                              [param.name for param in obj.parameters]...), 
+                         [Expr(:call, :convert, :(Variable{$(fieldparams[i])}), fields[i]) 
+                          for i in eachindex(fields)]...)
+
+    Expr(:(=), definition_ls, definition_rs)
 end
     
-
+macro outer_constructor(obj, defaults)
+    eval(generate_outer_constructor(eval(obj), eval(defaults)))
+end
 
 #=
     targettypes = [i == findfirst(fieldtypes, fieldtype) ?